+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - will not internally change commands starting with a reserved keyword (ex: `if s ==` will not be changed to `s = if s ===`)
 - generator
 - skip processing if not needed to speed up
-- `--format` flag
+- `--format`, `--stderr` flags
 
 ## 0.9 (2020-12-02)
 - other modules added for auto-import